refactor(visualization): migrate viewCube to a feature slice
Story #2, PR 4 of 5. Move ui/viewCube into features/viewCube (the last ui/
folder) and finish the feature-architecture migration of the UI.

- Move the cube component + meshGenerator/materials/mouseEvents service +
  zoomSlider into features/viewCube; add facade.ts exposing ViewCubeComponent
  and ViewCubeMouseEventsService. viewCube has no @ngrx/store usage, so no
  stores are needed. Repoint codeMap's three viewCube imports to the facade
  (codeMap <-> viewCube is a cross-feature cycle, allowed as a warning).
- Convert the 2 SCSS files to Tailwind: viewCube host (absolute/right/
  transition + [&.sideBarVisible]:right-[var(--cc-inspector-width)]) and the
  zoomSlider (vertical range input incl. ::-webkit/-moz thumb via arbitrary
  variants); tests key off data-testid, not classes. ui/ is now empty.

Fix a production-build regression introduced in PR 3: the codeMap facade
re-exported CodeMapPageObject from codeMap.po, which esbuild does not
tree-shake, pulling codeMap.po -> playwright.helper -> node "path" into the
browser bundle (ng build failed with `Could not resolve "path"`). Drop the
page-object re-export, import it directly in the one cross-feature e2e, and
exempt .e2e.ts/.po.ts from feature-cross-feature-only-via-public-api.

Behavior unchanged. tsc, full Jest suite (378 suites), dep-cruiser (0 errors)
and `ng build` (exit 0) pass.
