+++ visualization/app/codeCharta/features/codeMap/facade.ts
@@ -9,4 +9,3 @@ export { ThreeCameraService } from "./threeViewer/threeCamera.service"
 export { CodeMapBuilding } from "./rendering/codeMapBuilding"
 export { CodeMapMesh } from "./rendering/codeMapMesh"
 export { indicesPerNode } from "./rendering/geometryGenerationHelper"
-export { CodeMapPageObject } from "./codeMap.po"
