+++ visualization/app/codeCharta/features/viewCube/facade.ts
@@ -1,0 +1,2 @@
+export { ViewCubeComponent } from "./viewCube.component"
+export { ViewCubeMouseEventsService } from "./viewCube.mouseEvents.service"
