+++ visualization/app/codeCharta/features/viewCube/zoomSlider/zoomSlider.component.spec.ts
