--- visualization/app/codeCharta/ui/viewCube/viewCube.component.scss
@@ -1,11 +1,0 @@
-:host {
-    transition: right 0.3s ease;
-    position: absolute;
-    z-index: 11;
-    right: 0;
-    top: 10px;
-
-    &.sideBarVisible {
-        right: var(--cc-inspector-width);
-    }
-}
