--- visualization/app/codeCharta/ui/viewCube/zoomSlider/zoomSlider.component.html
@@ -1,15 +1,0 @@
-<div class="zoom-slider">
-    <button class="zoom-button" data-testid="zoomIn" (click)="zoomIn()">+</button>
-    <input
-        type="range"
-        id="zoomRange"
-        data-testid="zoomRange"
-        aria-label="Zoom level"
-        [min]="minZoom"
-        [max]="maxZoom"
-        [value]="zoomPercentage"
-        (input)="onInput($event)"
-    />
-    <button class="zoom-button" data-testid="zoomOut" (click)="zoomOut()">-</button>
-    <span id="zoomValue">{{ zoomPercentage?.toFixed(0) }}%</span>
-</div>
