--- visualization/app/codeCharta/ui/viewCube/zoomSlider/zoomSlider.component.scss
@@ -1,62 +1,0 @@
-$base-color: rgb(115, 115, 115);
-
-.zoom-slider {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    gap: 10px;
-    height: 150px;
-    position: absolute;
-    top: 10px;
-    right: 5px;
-    margin: 10px;
-
-    .zoom-button,
-    #zoomValue {
-        font-size: 10px;
-        color: $base-color;
-
-        &.zoom-button {
-            background: none;
-            border: none;
-            font-size: 16px;
-            cursor: pointer;
-            padding: 0;
-            margin: 0;
-        }
-    }
-
-    #zoomValue {
-        width: 30px;
-        text-align: center;
-    }
-}
-
-input[type="range"] {
-    -webkit-appearance: none;
-    writing-mode: vertical-lr;
-    direction: rtl;
-    color: $base-color;
-    width: 3px;
-    height: 100px;
-    background: $base-color;
-    outline: none;
-    opacity: 0.7;
-    transition: opacity 0.2s;
-
-    &::-webkit-slider-thumb {
-        -webkit-appearance: none;
-        appearance: none;
-        width: 10px;
-        height: 3px;
-        background: $base-color;
-        cursor: pointer;
-    }
-
-    &::-moz-range-thumb {
-        width: 10px;
-        height: 3px;
-        background: $base-color;
-        cursor: pointer;
-    }
-}
